HID: multitouch: Disable event reporting on suspend (WIP)
Disable event reporting on suspend. This fixes spurious wakeups
on an Asus T101HA when the lid is closed. Without this change the touchpad
on this model will send events when the display is close to it, causing
immediate wakeup when the lid is closed.

Disabling event reporting should also help HID-multitouch devices save
power when suspened.

WIP:

This is marked as WIP / not ready for merging because this also fully
disables wakeup by e.g. moving a finger over a HID-multitouch touchpad,
which is not necessarily desirable.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>