pinctrl: cherryview: Enable glitch filter for GPIOs used as interrupts
On Bay Trail we enable glitch filtering for GPIOs used as interrupts by
default to avoid spurious interrupts. So far this does not really seem
to be an issue on Cherry Trail devices.

With that said enabling glitch-filtering for pins used as interrupts
should not cause any issues and probably is a good idea regardless.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>